refactor(core,daisyui): put registry assembly in Theme.ts, drop daisyui's no-op fields
buildRegistry() and its getGlobalFormOptions() helper describe the core theme
plus a form's overrides, which is Theme.ts's subject, not Form.tsx's. The
FormProps dependency is type-only, so the resulting cycle is erased.

generateButtonTemplates() becomes a named export like its three siblings,
instead of a default renamed at the import site.

@rjsf/daisyui's generateTheme() returned the unmodified core fields, which
buildRegistry() already merges underneath a theme's overrides. No other theme
passes fields.
