+++ packages/core/src/components/Form.tsx
@@ -21,11 +21,9 @@ import type {
   ValidatorType,
   DefaultFormStateBehavior,
   CustomMergeAllOf,
-  GlobalFormOptions,
   NameGeneratorFunction,
 } from '@rjsf/utils';
 import {
-  englishStringTranslator,
   getByPath,
   setByPath,
   toPath,
@@ -43,19 +41,15 @@ import {
   SUBMIT_BTN_OPTIONS_KEY,
   toErrorList,
   toFieldPathId,
-  UI_DEFINITIONS_KEY,
-  UI_GLOBAL_OPTIONS_KEY,
   UI_OPTIONS_KEY,
   validationDataMerge,
-  DEFAULT_ID_SEPARATOR,
-  DEFAULT_ID_PREFIX,
   ERRORS_KEY,
   ID_KEY,
   ANY_OF_KEY,
   ONE_OF_KEY,
 } from '@rjsf/utils';
 
-import { generateTheme } from '../Theme.ts';
+import { buildRegistry } from '../Theme.ts';
 import { ADDITIONAL_PROPERTY_KEY_REMOVE, IS_RESET } from './constants.ts';
 
 /** The properties that are passed to the `Form` */
@@ -1395,65 +1389,4 @@ export default class Form<
       </FormTag>
     );
   }
-}
-
-/** Extracts the `GlobalFormOptions` from the given Form `props`
- *
- * @param props - The form props to extract the global form options from
- * @returns - The `GlobalFormOptions` from the props
- */
-function getGlobalFormOptions<T = any, S extends StrictRJSFSchema = RJSFSchema, F extends FormContextType = any>(
-  props: FormProps<T, S, F>,
-): GlobalFormOptions {
-  const {
-    experimental_componentUpdateStrategy,
-    idSeparator = DEFAULT_ID_SEPARATOR,
-    idPrefix = DEFAULT_ID_PREFIX,
-    nameGenerator,
-    useFallbackUiForUnsupportedType = false,
-  } = props;
-  // Omit any options that are undefined or null
-  return {
-    idPrefix,
-    idSeparator,
-    useFallbackUiForUnsupportedType,
-    ...(experimental_componentUpdateStrategy !== undefined && { experimental_componentUpdateStrategy }),
-    ...(nameGenerator !== undefined && { nameGenerator }),
-  };
-}
-
-/** Builds the `Registry` for a form from the given `props`, `schema` and `schemaUtils`
- *
- * @param props - The form props to build the registry from
- * @param schema - The root schema for the form
- * @param schemaUtils - The `SchemaUtilsType` implementation to put on the registry
- * @returns - The `Registry` described by the inputs
- */
-export function buildRegistry<T = any, S extends StrictRJSFSchema = RJSFSchema, F extends FormContextType = any>(
-  props: FormProps<T, S, F>,
-  schema: S,
-  schemaUtils: SchemaUtilsType<T, S, F>,
-): Registry<T, S, F> {
-  const { translateString = englishStringTranslator, uiSchema = {} } = props;
-  const { fields, templates, widgets } = generateTheme<T, S, F>();
-  return {
-    fields: { ...fields, ...props.fields },
-    templates: {
-      ...templates,
-      ...props.templates,
-      ButtonTemplates: {
-        ...templates.ButtonTemplates,
-        ...props.templates?.ButtonTemplates,
-      },
-    },
-    widgets: { ...widgets, ...props.widgets },
-    rootSchema: schema,
-    // `F` may be narrower than `{}`; an omitted formContext has always defaulted to an empty object regardless
-    formContext: props.formContext ?? ({} as F),
-    schemaUtils,
-    translateString,
-    globalUiOptions: uiSchema[UI_GLOBAL_OPTIONS_KEY],
-    globalFormOptions: getGlobalFormOptions(props),
-    uiSchemaDefinitions: uiSchema[UI_DEFINITIONS_KEY] ?? {},
-  };
 }
